+++ bootstrap.ml
@@ -36,6 +36,7 @@ type subdirs =
 let dirs =
   [ ("src/stdune/result", Some "Dune_result", No)
   ; ("src/stdune/caml", Some "Dune_caml", No)
+  ; ("src/dune_action_plugin", Some "Dune_action_plugin", No)
   ; ("src/stdune", Some "Stdune", No)
   ; ("src/fiber", Some "Fiber", No)
   ; ("src/xdg", Some "Xdg", No)