+++ src/dune/action_dune_lang.mli
@@ -11,6 +11,9 @@ include
 
 include Dune_lang.Conv.S with type t := t
 
+(** Raises User_error on invalid action. *)
+val validate : loc:Loc.t -> t -> unit
+
 include
   Action_intf.Helpers
     with type t := t