+++ src/dune/build.ml
@@ -263,7 +263,7 @@ let static_deps t ~list_targets =
       acc
     | Deps deps -> Static_deps.add_action_deps acc deps
     | Paths_for_rule fns -> Static_deps.add_rule_paths acc fns
-    | Paths_glob g -> Static_deps.add_action_dep acc (Dep.glob g)
+    | Paths_glob g -> Static_deps.add_action_dep acc (Dep.file_selector g)
     | If_file_exists (p, state) -> (
       match !state with
       | Decided (_, t) -> loop t acc false