+++ src/dune/glob.ml
@@ -6,6 +6,8 @@ type t =
   ; repr : string
   }
 
+let compare x y = String.compare x.repr y.repr
+
 let equal x y = String.equal x.repr y.repr
 
 let hash t = String.hash t.repr
@@ -21,6 +23,10 @@ let of_string_exn loc repr =
   | Error (_, msg) -> User_error.raise ~loc [ Pp.textf "invalid glob: :%s" msg ]
   | Ok t -> t
 
+let encode t =
+  let open Dune_lang.Encoder in
+  string t.repr
+
 let decode =
   let open Dune_lang.Decoder in
   plain_string (fun ~loc str -> of_string_exn loc str)
@@ -30,6 +36,8 @@ let test t = Re.execp t.re
 let filter t = List.filter ~f:(test t)
 
 let empty = { re = Re.compile Re.empty; repr = "\000" }
+
+let universal = { re = Re.compile (Re.rep Re.any); repr = "**" }
 
 let to_pred t =
   let id =