+++ src/dune/glob.mli
@@ -4,9 +4,13 @@ type t
 
 val equal : t -> t -> bool
 
+val compare : t -> t -> Ordering.t
+
 val hash : t -> int
 
 val to_dyn : t Dyn.Encoder.t
+
+val encode : t Dune_lang.Encoder.t
 
 val decode : t Dune_lang.Decoder.t
 
@@ -15,6 +19,8 @@ val test : t -> string -> bool
 val filter : t -> string list -> string list
 
 val empty : t
+
+val universal : t
 
 val of_string_exn : Loc.t -> string -> t