+++ test/blackbox-tests/test-cases-with-libs/dune-action-plugin/multiple-dynamic-run-within-single-action/bin/foo.ml
@@ -1,0 +1,3 @@
+open Dune_action_plugin.V1
+
+let () = run (return ())