+++ test/blackbox-tests/test-cases-with-libs/dune-action-plugin/no-dependencies/test/run.t
@@ -1,0 +1,16 @@
+This test checks that executable that uses 'dynamic-run'
+but requires no dependencies can be successfully run.
+
+  $ echo "(lang dune 2.0)" > dune-project
+
+  $ cat > dune << EOF
+  > (alias
+  >  (name runtest)
+  >  (action (dynamic-run ./foo.exe)))
+  > EOF
+
+  $ cp ../bin/foo.exe ./
+
+  $ dune runtest --display short
+           foo alias runtest
+  Hello from foo!